update to relocated packages
